(authfeat): 添加用户认证和权限控制功能
- 新增用户登录JWT的用户认证和权限控制和注册接口
- 实现基于- 添加全局异常处理
-优化用户和角色相关接口
- 更新数据库表结构，将用户ID改为Long类型
- 移除跨域配置类，改为在SecurityConfig中配置跨域

+++ src/main/java/com/example/demo/service/UserService.java
@@ -24,7 +24,7 @@ public interface UserService extends IService<User> {
     void updateUserRoleService(User user);
 
     /*实现用户信息的删除*/
-    void deleteUserRoleService(Integer id);
+    void deleteUserRoleService(Long id);
 
     /*实现员工按照年龄段分布统计*/
     List<CountResult> countEmployeeAageService();
